+++ traffic_signaling/include/simulation/street.hpp
@@ -12,22 +12,22 @@ public:
     explicit Street(const city_plan::Street &data)
         : data_(data) {}
 
-    void add_car(size_t car_id, size_t time) {
+    void add_car(unsigned long car_id, unsigned long time) {
         car_queue_.emplace(car_id);
         latest_used_time_ = time;
     }
 
-    size_t get_car() {
+    unsigned long get_car() {
         auto &&car = car_queue_.front();
         car_queue_.pop();
         return car;
     }
 
-    size_t id() const {
+    unsigned long id() const {
         return data_.id();
     }
 
-    std::optional<size_t> latest_used_time() const {
+    std::optional<unsigned long> latest_used_time() const {
         return latest_used_time_;
     }
 
@@ -40,8 +40,8 @@ private:
     const city_plan::Street &data_;
 
     // car queue represented by car ids
-    std::queue<size_t> car_queue_;
-    std::optional<size_t> latest_used_time_;
+    std::queue<unsigned long> car_queue_;
+    std::optional<unsigned long> latest_used_time_;
 };
 }
